Issue #68, fixed how reports tallied in most frequent drug table
Multiple drugs of the same name may be listed as suspect/concomitant for the same REPORT_ID if they are listed with different dosages or forms, leading to discrepancy in number of reports. Added a filter before this step. Also, excluded reports with empty string as reaction, as this led to a discrepancy in total reports count.